+++ src/jsii-project.ts
@@ -260,7 +260,7 @@ export class JsiiProject extends TypeScriptProject {
 
     const compat = options.compat ?? false;
     if (compat) {
-      this.addCompileCommand('yarn compat');
+      this.addCompileCommand(`${this.runScriptCommand} compat`);
     } else {
       this.addTip('Set "compat" to "true" to enable automatic API breaking-change validation');
     }